implementing dataset api network from binar academy

+++ app/src/main/java/com/lynn/foodies/data/repository/CartRepository.kt
@@ -88,7 +88,7 @@ class CartRepositoryImpl(private val cartDataSource: CartDataSource) : CartRepos
                         itemQuantity = quantity,
                         catalogName = catalog.name,
                         catalogImageUrl = catalog.imageUrl,
-                        catalogPrice = catalog.price,
+                        catalogPrice = catalog.price.toDouble(),
                         itemNotes = notes
                     )
                 )
