+++ app/src/main/java/com/lynn/foodies/presentation/home/adapter/CatalogListItemViewHolder.kt
@@ -17,7 +17,7 @@ class CatalogListItemViewHolder(
                 crossfade(true)
             }
             binding.tvCatalogName.text = it.name
-            binding.tvCatalogPrice.text = it.price.toIndonesianFormat()
+            binding.tvCatalogPrice.text = it.price.toDouble().toIndonesianFormat()
             itemView.setOnClickListener {
                 listener.onItemClicked(item)
             }
